% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{density_folded_normal}
\alias{density_folded_normal}
\title{Folded normal density}
\usage{
density_folded_normal(x, mu, sigma)
}
\arguments{
\item{x}{Numeric vector of observations.}

\item{mu}{Mean of the underlying normal distribution.}

\item{sigma}{Standard deviation (> 0) of the underlying normal distribution.}
}
\value{
Numeric vector of densities.
}
\description{
Computes the density of the folded normal distribution for `x` given
parameters `mu` and `sigma`. Values of `x` are treated as absolute values.
}
\keyword{internal}
