% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Retrieve compiled Stan model, using cached copy if available}
\usage{
get_stan_model()
}
\value{
A compiled Stan model object.
}
\description{
This function looks for a previously cached compiled Stan model in a
user cache directory. If found, it loads and returns it. If a model has
already been loaded in this session, it returns that. If no cached model
is found, it tries to locate a packaged model file via \code{system.file}.
If none is found, an error is raised instructing the user to compile
the Stan code once via \code{compile_stan_model(stan_code)}.
}
